inject room to project

+++ FoodDeliveryAndroid/app/src/main/java/com/example/fooddeliveryandroid/data/remote/dto/CartItemResponse.kt
@@ -2,5 +2,5 @@ package com.example.fooddeliveryandroid.data.remote.dto
 
 data class CartItemResponse (
     val product: ProductCartItemResponse,
-    val quantity: Short,
+    val quantity: Int,
 )
