+++ FoodDeliveryAndroid/app/src/main/java/com/example/fooddeliveryandroid/data/remote/dto/request/UpdateCartItemQuantityRequest.kt
@@ -1,5 +1,5 @@
 package com.example.fooddeliveryandroid.data.remote.dto.request
 
 data class UpdateCartItemQuantityRequest (
-    val quantity: Short
+    val quantity: Int
 )
